% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_normalize_label_data.R
\name{add_neuron_id}
\alias{add_neuron_id}
\title{Add Neuron Identifier as the Experiment Name and Neuron Index}
\usage{
add_neuron_id(exp, exp_name)
}
\arguments{
\item{exp}{\link{experiment}, a \link{meapr} experiment object}

\item{exp_name}{\code{character} name for the experiment}
}
\value{
The \link{experiment} that was passed in as \code{exp}, with two
additional column, \code{exp_id} with the experiment name given in \code{exp_name}
and \code{neuron_id} with the \verb{<exp_name>_<neuron_index>}
}
\description{
Add Neuron Identifier as the Experiment Name and Neuron Index
}
