% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-spike-features.R
\name{compute_spike_features}
\alias{compute_spike_features}
\title{Compute Spike Features}
\usage{
compute_spike_features(spike_times, seg_length = 30, burst_labels = NULL)
}
\arguments{
\item{spike_times}{\code{data.frame} with columns
\code{[neuron_index, time_step, treatment, begin, end]} and a row for each
detected firing, e.g. from \code{load_experiment_phy()$firing}}

\item{seg_length}{integer number of seconds to wrap on to each line of the
resulting plot}

\item{burst_labels}{\code{vector} results of \code{compute_bursts()}, and gives which
spikes belong in a burst}
}
\value{
a \code{list} with elements of \code{rats}
}
\description{
A spike is an individual or averaged electrical signal captured on electrodes
the shape depends on the neuron action potential generating the spike, and
the orientation of the axon and resulting electric field relative to the
electrode. This function computes key features of the spike that can be
used to assess neuron type and treatment effects.
}
