% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{date_code}
\alias{date_code}
\title{create date code of the form YYMMDD}
\usage{
date_code(d = NA)
}
\arguments{
\item{d}{\code{data.code} in the format of \code{Sys.Date()} for which to generate the
date code, defaulting to 'today'.}
}
\description{
create date code of the form YYMMDD
}
