% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_normalize_label_data.R
\name{exposure_counts}
\alias{exposure_counts}
\title{Compute the Firing Rate for Each Treatment}
\usage{
exposure_counts(exp)
}
\arguments{
\item{exp}{\link{experiment}, a \link{meapr} experiment object}
}
\value{
\code{data.frame} with one treatment per row, and columns
\itemize{
\item \strong{begin}: beginning time of the treatment (in seconds)
\item \strong{end}: end time of the treatment (in seconds)
\item \strong{count}: number of firing events during the treatment
\item \strong{exposure}: time duration of the treatment (in seconds)
\item \strong{firing_rate}: average number of firing evens per second
}
}
\description{
Compute the Firing Rate for Each Treatment
}
