% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-waveform-features.R
\name{extract_waveform_features}
\alias{extract_waveform_features}
\title{Extract Waveform Features}
\usage{
extract_waveform_features(wf)
}
\arguments{
\item{wf}{\code{vector} of the waveform amplitudes over timesteps}
}
\value{
list of waveform features with fields
\itemize{
\item file_name
\item unit_num
\item file_index
\item class
\item mainLoc
\item locs
\item bpd
\item bpd_t
\item p2p
\item p2p_t
\item mainAmp
\item secAmp
\item peak2Trough
\item width
\item width_t
\item halfWidth
\item halfWidth_t
}
}
\description{
Extract Waveform Features
}
