% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-experiment-matlab.R
\name{load_experiment_matlab}
\alias{load_experiment_matlab}
\title{Load Firing Data for an Experiment from MATLAB}
\usage{
load_experiment_matlab(
  units_fname,
  treatments,
  experiment_tag = NULL,
  time_steps_per_second = 1,
  save_path = "intermediate_data/experiment_datasets",
  verbose = FALSE
)
}
\arguments{
\item{units_fname}{\code{character}. A file name to a should be a \code{.mat} file
exported from the Margolis lab multi-electrode experimental MATLAB
processing pipeline containing the following information for each neuron
that was detected
\itemize{
\item{\strong{1) }}{The time steps (in seconds) when it fired during the
experiment}
\item{\strong{2) }}{The waveform of the cluster center across all
firings}
}}

\item{treatments}{\code{data.frame} or \code{character}. If a \code{data.frame} it should
have columns \verb{[treatment, begin, end]} for each treatment in the
experiment, where begin and end are given as seconds since the beginning of
the experiment. If it is a \code{character} it should be a path to a \code{.csv} file
with the same columns. To help detect problems, an warning is given if the
treatments are not disjoint and given chronologically.}

\item{experiment_tag}{\code{character} an identifier for the experiment, set
in the return data structure and path to save to disk if null (default),
then use \code{treatments |> basename() |> stringr::str_replace(".mat$",
"")}}

\item{time_steps_per_second}{\code{numeric} scaling factor to convert time steps
to seconds. E.g. if the recording is done at 20k Hz, then use \code{20000}.}

\item{save_path}{\code{character} file path where loaded data set should be
cached: \verb{<save_path>/<experiment_tag>}}

\item{verbose}{\code{logical} print out verbose output.}
}
\value{
\link{experiment} S3 class with the following elements
\itemize{
\item{\strong{tag: }}{<experiment_tag>}
\item{\strong{treatment: }}{\code{\link[tibble]{tibble}} with columns
\code{[treatment, begin, end]}}
\item{\strong{firings: }}{\code{\link[tibble]{tibble}} with columns
\code{[neuron_index, time_step, treatment, begin, end]} and a row for
each detected firing}
\item{\strong{waveform: }}{\code{\link[tibble]{tibble}} with columns
\code{[neuron_index, time_step, voltage]} for each neuron}
}
}
\description{
Load Firing Data for an Experiment from MATLAB
}
