% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-experiment-phy.R
\name{load_experiment_phy}
\alias{load_experiment_phy}
\title{Load Firing Data for an Experiment from phy}
\usage{
load_experiment_phy(
  data_path,
  treatments,
  experiment_tag = NULL,
  time_steps_per_second = 20000,
  save_path = "intermediate_data/experiment_datasets",
  verbose = FALSE
)
}
\arguments{
\item{data_path}{\code{character} path to phy data path. It should be
a directory with the output of running phy spike sorter, usually
with an extension of \code{.modat.GUI}.}

\item{treatments}{\code{data.frame} or \code{character}. See \link{load_treatments_file}
for details.}

\item{experiment_tag}{\code{character} an identifier for the experiment, set
in the return data structure and path to save to disk.}

\item{time_steps_per_second}{numeric number of time steps per second.}

\item{save_path}{\code{character} file path where loaded data set should be
cached: \verb{<save_path>/<experiment_tag>}}

\item{verbose}{\code{logical} print out verbose output}
}
\value{
\link{meapr-experiment} S3 class with the following elements
\itemize{
\item{\strong{tag: }}{<experiment_tag>}
\item{\strong{treatment: }}{\code{\link[tibble]{tibble}} with columns
\code{[treatment, begin, end]}} where treatment is a treatment
identifier, and begin and end are the time points in seconds of the
beginning and ending of the treatment.
\item{\strong{firings: }}{\code{\link[tibble]{tibble}} with columns
\code{[neuron_index, time_step, treatment, begin, end]} and a row for
each detected firing}
\item{\strong{waveform: }}{\code{\link[tibble]{tibble}} with columns
\code{[neuron_index, time_step, voltage]} for each neuron}
}
}
\description{
(Phy)\link{https://github.com/cortex-lab/phy} is a data exchange
framework for electrophysiology data. For example, the SpyKING Circus spike
sorter data can be exported in phy format. This function will load data from
the export in the \verb{*.modat.GUI/} folder

The data is loaded into a list that has fields [tag, treatment, firings,
waveform].
}
