% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-experiment-spyking_circus.R
\name{load_experiment_spiking_circus}
\alias{load_experiment_spiking_circus}
\title{Load Firing Data for an Experiment from SpyKING Circus}
\usage{
load_experiment_spiking_circus(
  data_path,
  treatments,
  experiment_tag = NULL,
  save_path = "intermediate_data/experiment_datasets",
  verbose = FALSE
)
}
\arguments{
\item{data_path}{\code{character} path to SpyKING Circus data path. It should be
a directory with the output of running SpyKING Circus spike sorter, usually
with an extension of \code{.modat.GUI}.}

\item{treatments}{\code{data.frame} or \code{character}. See \link{load_treatments_file}
for details.}

\item{experiment_tag}{\code{character} an identifier for the experiment, set
in the return data structure and path to save to disk.}

\item{save_path}{\code{character} file path where loaded data set should be
cached: \verb{<save_path>/<experiment_tag>}}

\item{verbose}{\code{logical} print out verbose output}
}
\value{
\link{meapr-experiment} S3 class with the following elements
\itemize{
\item{\strong{tag: }}{<experiment_tag>}
\item{\strong{treatment: }}{\code{\link[tibble]{tibble}} with columns
\code{[treatment, begin, end]}}
\item{\strong{firings: }}{\code{\link[tibble]{tibble}} with columns
\code{[neuron_index, time_step, treatment, begin, end]} and a row for
each detected firing}
\item{\strong{waveform: }}{\code{\link[tibble]{tibble}} with columns
\code{[neuron_index, time_step, voltage]} for each neuron}
}
}
\description{
Load Firing Data for an Experiment from SpyKING Circus
}
