% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-recording.R
\name{load_recording}
\alias{load_recording}
\title{Load MED64 Recording Trace}
\usage{
load_recording(data_path, n_channels = 64, verbose = FALSE)
}
\arguments{
\item{data_path}{\code{character} path to \code{modat.bin} file}

\item{n_channels}{\code{integer} number of channels (Default: 64)}

\item{verbose}{\code{logical} verbose output (Default: FALSE)}
}
\value{
\code{matrix} with dimensions \verb{[n_channels, n_samples]} with values
as signed integers, typically in the range e.g. \verb{[-100, 100]}.
}
\description{
Load the raw recording trace in as an array. This assumes the
data is organized as raw signed 16bit integers as a little endian
column-major array.
}
\examples{
\dontrun{
  recording <- meapr::load_recording(
    data_path = '20250131_15h42m41s.modat.bin')
  shape(recording)
}

\dontrun{
  extract_path <- tempdir()
  unzip(
    zipfile = "20250131_15h42m41s.modat.bin.filtered.zip",
    extdir = extract_path)
  recording <- meapr::load_recording(
    data_path = paste0(extract_path, "/20250131_15h42m41s.modat.bin"))
}

}
