% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-treatments-file.R
\name{load_treatments_file}
\alias{load_treatments_file}
\title{Load treatments file}
\usage{
load_treatments_file(treatments, verbose = FALSE)
}
\arguments{
\item{treatments}{\code{data.frame} or \code{character}. If a \code{data.frame} it should
have columns \verb{[index, treatment, begin, end]} for each treatment in the
experiment, where begin and end are given as seconds since the beginning
of the experiment. If it is a \code{character} it should be a path to a \code{.tsv}
file with the same columns. To help detect problems, an warning is given
if the treatments are not disjoint and given chronologically.}

\item{verbose}{\code{logical} print out verbose output}
}
\value{
\code{data.frame} with treatment information. See Description for the
format.
}
\description{
A treatments data table has columns \code{begin}, and \code{end}, and
\code{treatment}, where \code{begin} and \code{end} are the time-points in seconds when
the treatment is applied.
}
