% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_normalize_label_data.R
\name{norm_firing}
\alias{norm_firing}
\title{For Each Neuron, Normalize Firing Rate to Percent Baseline}
\usage{
norm_firing(exp, baseline_treatment_name = "Baseline")
}
\arguments{
\item{exp}{\link{experiment}, a \link{meapr} experiment object}

\item{baseline_treatment_name}{\code{character} name of the baseline treatment}
}
\value{
The \link{experiment} that was passed in as \code{exp}, add an
additional column \code{norm_firing_rate} that is the percent of the firing rate
in the Baseline condition
}
\description{
For Each Neuron, Normalize Firing Rate to Percent Baseline
}
