% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-treatments-file.R
\name{parse_concentration}
\alias{parse_concentration}
\title{Parse Concentration}
\usage{
parse_concentration(concentration_string)
}
\arguments{
\item{concentration_string}{string number formatted as
"\if{html}{\out{<numeric>}} \if{html}{\out{<SI units>}}", where \if{html}{\out{<SI units>}} is one of
[M, mM, μM, uM, nM, pM, fM]}
}
\value{
numeric molar concentration
}
\description{
Parse concentration with SI units into molar units
}
\examples{
\dontrun{
  1e-8 == parse_concentration("10 nM")
  1.3e-6 == parse_concentration("1.3 uM")
}

}
