% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_marginal_effects.R
\name{plot_model_marginal_effects}
\alias{plot_model_marginal_effects}
\title{Plot Model Conditional Effects}
\usage{
plot_model_marginal_effects(
  model_fit,
  model_tag,
  plot_width = 6,
  plot_height = 6,
  output_base = "product/plots",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{model_fit}{\link[brms:brmsfit-class]{brms::brmsfit} model object}

\item{plot_width}{\code{numeric} width of the output plot}

\item{plot_height}{\code{numeric} height of the output plot}

\item{verbose}{\code{logical} print out verbose output}
}
\value{
: \link[ggplot2:ggplot]{ggplot2::ggplot} of the plot and it saves the result to
\verb{product/plots/firing_qqplot_by_treatment_<experiment_tag>_<date_code>.(pdf|png)}
It save both .pdf and .png because it's easier to email etc small pngs
while for use in an a manuscript having the vector version means that it`
can be tweaked with illustrator
}
\description{
Plot Model Conditional Effects
}
