% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recording_amplitude_histogram.R
\name{plot_recording_amplitude_density}
\alias{plot_recording_amplitude_density}
\title{Plot Density of Recording Amplitudes by Channel}
\usage{
plot_recording_amplitude_density(
  recording,
  extra_layers,
  width = 10,
  height = 10,
  output_base,
  verbose = FALSE
)
}
\arguments{
\item{recording}{\code{matrix} result of \code{meap::load_recording()}:
a \code{matrix} with dimensions \verb{[n_channels, n_samples]} with values as signed
integers, typically in the range e.g. \verb{[-100, 100]}.}

\item{extra_layers}{\code{list} extra ggplot2 layers to be added to the plot
before saving it.}

\item{output_base}{\code{character} where to output plots}

\item{verbose}{\code{logical} print out verbose output}

\item{plot_width}{\code{numeric} width of the output plot}

\item{plot_height}{\code{numeric} height of the output plot}
}
\value{
: \link[ggplot2:ggplot]{ggplot2::ggplot} of the plot and it saves the result to
\verb{product/plots/firing_qqplot_by_treatment_<experiment_tag>_<date_code>.(pdf|png)}
It save both .pdf and .png because it's easier to email etc small pngs
while for use in an a manuscript having the vector version means that it`
can be tweaked with illustrator
}
\description{
Plot Density of Recording Amplitudes by Channel
}
