% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_run_apps.R
\name{shiny_run_curate_units}
\alias{shiny_run_curate_units}
\title{Run the Curate Units Shiny Application}
\usage{
shiny_run_curate_units()
}
\description{
Usage:
}
\details{
\enumerate{
\item Select a loaded data set
\item Unit --> Firing pattern
\itemize{
\item Select a unit from a list or small-multiples plot
\item hilight firing rate of unit in plots
\itemize{
\item Average firing rate of unit across treatments
\item Firing density across the experiment duration
\item Firing qq-plot across treatments
}
}
\item Brushing Firing Patterns
\itemize{
\item Select query/response plots
\item Select unit in query plot
\item highlight unit in response plot
}
}
}
