% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_treatments_add_log_dose.R
\name{treatment_log_dose}
\alias{treatment_log_dose}
\title{Filter to the Treatment Condition and the log10(treatment dose)}
\usage{
treatment_log_dose(labeled_data)
}
\arguments{
\item{labeled_data}{\code{data.frame}}
}
\description{
Filter to the Treatment Condition and the log10(treatment dose)
}
