% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_normalize_label_data.R
\name{filter_data_baseline}
\alias{filter_data_baseline}
\title{Filter Neurons by the Firing Rate in the Baseline}
\usage{
filter_data_baseline(exp, lower, upper, baseline_treatment_name = "Baseline")
}
\arguments{
\item{exp}{\link{meapr-experiment}, a \link{meapr} experiment object}

\item{lower}{\code{numeric} lower firing rate to filter by (inclusive)}

\item{upper}{\code{numeric} upper firing rate to filter by (inclusive)}

\item{baseline_treatment_name}{\code{character} name of the baseline treatment}
}
\value{
The \link{meapr-experiment} that was passed in as \code{exp}, but with only
the neurons where the firing rate is within
\verb{lower <= <firing_rate> <= upper}
}
\description{
Filter Neurons by the Firing Rate in the Baseline
}
