% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_normalize_label_data.R
\name{label_Hz}
\alias{label_Hz}
\title{Label Neurons That Have a Firing Rate of At least 0.5 Hz in the Baseline}
\usage{
label_Hz(exp, baseline_treatment_name = "Baseline")
}
\arguments{
\item{exp}{\link{meapr-experiment}, a \link{meapr} experiment object}

\item{baseline_treatment_name}{\code{character} name of the baseline treatment}
}
\value{
The \link{meapr-experiment} that was passed in as \code{exp}, with an
additional column \code{Hz} with values \code{"< 0.5 Hz"} or \code{">= 0.5 Hz"} depending
on the firing rate in the given baseline condition.
}
\description{
Label Neurons That Have a Firing Rate of At least 0.5 Hz in the Baseline
}
