% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_normalize_label_data.R
\name{label_responsiveness}
\alias{label_responsiveness}
\title{Label Each Neuron by it Responsiveness Relative to a Threshold Range}
\usage{
label_responsiveness(exp, min_threshold, max_threshold)
}
\arguments{
\item{exp}{\link{meapr-experiment}, a \link{meapr} experiment object}

\item{min_threshold}{\code{numeric} lower firing rate threshold for responsiveness}

\item{max_threshold}{\code{numeric} upper firing rate threshold for responsiveness}
}
\value{
The \link{meapr-experiment} that was passed in as \code{exp}, with an
additional column \code{responsive} taking values
\verb{["decreasing", "increasing", "low_response"]}, depending on whether the
the where the normalized firing rate for the treatment condition below the
\code{min_threshold}, \code{rises} above the \code{max_threshold} or stays between them
}
\description{
Label Each Neuron by it Responsiveness Relative to a Threshold Range
}
