% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dose_response.R,
%   R/model_treatment_log_poisson.R
\name{model_treatment_log_poisson}
\alias{model_treatment_log_poisson}
\title{Model Average Firing Rate as a Function of Treatment Dose using Log-Poisson
Regression}
\usage{
model_treatment_log_poisson(
  experiment,
  output_base = "intermediate_data/models",
  verbose = FALSE,
  ...
)

model_treatment_log_poisson(
  experiment,
  output_base = "intermediate_data/models",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{experiment}{\link{experiment} data set loaded with
\link{load_experiment_matlab} or \link{load_experiment_phy}}

\item{output_base}{\code{character} path where the model should be saved. The
output file is \verb{<output_base>/model_log_poisson_<experiment$tag>.stan}}

\item{verbose}{\code{logical} print out verbose output}

\item{...arguments}{are passed to \link[brms:brm]{brms::brm}}
}
\value{
\link[brms:brmsfit]{brms::brmsfit} object with the fit model

\link[brms:brmsfit]{brms::brmsfit} object with the fit model
}
\description{
Build a Bayesian hierarchical model for the firing rate
as a function of the treatment and other covariates. This uses log-poisson
regression which  directly estimates the number of firing events with
taking into account the exposure.

treatment as a fixed effect
=> treatment

average neuron firing rate as a random effect
=> (1|neuron_index)

assume a general linear model with log-poisson link

Build a Bayesian hierarchical model for the firing rate
as a function of the treatment and other covariates. This uses log-poisson
regression which  directly estimates the number of firing events with
taking into account the exposure.

treatment as a fixed effect
=> treatment

average neuron firing rate as a random effect
=> (1|neuron_index)

assume a general linear model with log-poisson link
}
