% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines_firing_rate_by_treatment.R
\name{plot_lines_firing_rate_by_treatment}
\alias{plot_lines_firing_rate_by_treatment}
\title{Dot/boxplot plot of average firing rate by treatment}
\usage{
plot_lines_firing_rate_by_treatment(
  experiment,
  plot_width = 6,
  plot_height = 6,
  output_base = "product/plots",
  verbose = FALSE
)
}
\arguments{
\item{experiment}{\link{meapr-experiment} data set loaded with
\link{load_experiment_matlab} or \link{load_experiment_spyking_circus}}

\item{plot_width}{\code{numeric} width of the output plot}

\item{plot_height}{\code{numeric} height of the output plot}

\item{verbose}{\code{logical} print out verbose output}
}
\value{
: \link[ggplot2:ggplot]{ggplot2::ggplot} of the plot and it saves the result to
\verb{product/plots/firing_qqplot_by_treatment_<experiment_tag>_<date_code>.(pdf|png)}
It save both .pdf and .png because it's easier to email etc small pngs
while for use in an a manuscript having the vector version means that it`
can be tweaked with illustrator
}
\description{
Make a plot where
the x-axis are the treatments and
the y-axis is the average firing rate on the log scale
show a box (and wisker) plot for each treatment summarizing across neurons in the experiment
overlay points for each neuron jittering horizontally to allow them to be seen more easily
}
