% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stable_baseline.R
\name{plot_stable_baseline}
\alias{plot_stable_baseline}
\title{Plot Baseline Firing density to evaluate the stability}
\usage{
plot_stable_baseline(
  experiment,
  baseline_treatment_name = "baseline",
  event_threshold = 200,
  extra_layers = list(),
  plot_width = 10,
  plot_height = 4,
  output_base = "product/plots",
  verbose = FALSE
)
}
\arguments{
\item{experiment}{\link{experiment} data set loaded with
\link{load_experiment_matlab} or \link{load_experiment_phy}}

\item{baseline_treatment_name}{\code{character} name of the baseline treatment}

\item{extra_layers}{\code{list} extra ggplot2 layers to be added to the plot
before saving it.}

\item{plot_width}{\code{numeric} width of the output plot.}

\item{plot_height}{\code{numeric} height of the output plot.}

\item{output_base}{\code{character} where to output plots.}

\item{verbose}{\code{logical} print out verbose output.}

\item{event_reshold}{\code{numeric} how many events should be used to set the
baseline threshold}
}
\value{
: \link[ggplot2:ggplot]{ggplot2::ggplot} of the plot and it saves the result to
\verb{product/plots/firing_qqplot_by_treatment_<experiment_tag>_<date_code>.pdf}
and
\verb{product/plots/firing_qqplot_by_treatment_<experiment_tag>_<date_code>.png}
It save both .pdf and .png because it's easier to email etc small pngs
while for use in an a manuscript having the vector version means that it`
can be tweaked with illustrator
}
\description{
The baseline firing rate can be used to evaluate whether or not
the unit is firing enough before the treatment to evaluate treatment
effects. Since baselines firing rates may not be stable over time, there
is a trade-off of collecting a baseline for a short time to be less biased
or for a long time to have less variance. This plot shows how as the time
is of the baseline treatment is shortened, how the estimate changes. Using
a minimum threshold for the number of firing events <event_threshold>,
units are classified as either keep/not-keep in the plot.
}
