% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stable_baseline.R
\name{plot_stable_baseline}
\alias{plot_stable_baseline}
\title{Plot Baseline Firing density to evaluate the stability}
\usage{
plot_stable_baseline(
  experiment,
  baseline_treatment_name = "baseline",
  plot_width = 10,
  plot_height = 4,
  output_base = "product/plots",
  verbose = FALSE
)
}
\arguments{
\item{experiment}{\link{meapr-experiment} data set loaded with
\link{load_experiment_matlab} or \link{load_experiment_spyking_circus}}

\item{baseline_treatment_name}{\code{character} name of the baseline treatment}

\item{plot_width}{\code{numeric} width of the output plot}

\item{plot_height}{\code{numeric} height of the output plot}

\item{verbose}{\code{logical} print out verbose output.}
}
\value{
: \link[ggplot2:ggplot]{ggplot2::ggplot} of the plot and it saves the result to
\verb{product/plots/firing_qqplot_by_treatment_<experiment_tag>_<date_code>.pdf}
and
\verb{product/plots/firing_qqplot_by_treatment_<experiment_tag>_<date_code>.png}
It save both .pdf and .png because it's easier to email etc small pngs
while for use in an a manuscript having the vector version means that it`
can be tweaked with illustrator
}
\description{
Plot Baseline Firing density to evaluate the stability
}
