% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waveform_correlation_matrix.R
\name{plot_waveform_correlation_matrix}
\alias{plot_waveform_correlation_matrix}
\title{plot Correlation matrix of waveforms}
\usage{
plot_waveform_correlation_matrix(
  experiment,
  plot_width = 10,
  plot_height = 10,
  output_base = "product/plots",
  verbose = FALSE
)
}
\arguments{
\item{experiment}{\link{meapr-experiment} data set loaded with
\link{load_experiment_matlab} or \link{load_experiment_spyking_circus}}

\item{plot_width}{\code{numeric} width of the output plot}

\item{plot_height}{\code{numeric} height of the output plot}

\item{verbose}{\code{logical} print out verbose output}
}
\value{
: \link[ggplot2:ggplot]{ggplot2::ggplot} of the plot and it saves the result to
\verb{product/plots/firing_qqplot_by_treatment_<experiment_tag>_<date_code>.(pdf|png)}
It save both .pdf and .png because it's easier to email etc small pngs
while for use in an a manuscript having the vector version means that it`
can be tweaked with illustrator
}
\description{
Plot a correlation matrix as a heatmap between all waveforms in the experiment
}
