% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waveform_lattice.R
\name{plot_waveform_lattice}
\alias{plot_waveform_lattice}
\title{Simple lattice plot of the waveforms}
\usage{
plot_waveform_lattice(
  experiment,
  plot_width = 10,
  plot_height = 10,
  output_base = "product/plots",
  verbose = FALSE
)
}
\arguments{
\item{experiment}{\link{meapr-experiment} data set loaded with
\link{load_experiment_matlab} or \link{load_experiment_spyking_circus}}

\item{plot_width}{\code{numeric} width of the output plot}

\item{plot_height}{\code{numeric} height of the output plot}

\item{verbose}{\code{logical} print out verbose output}
}
\value{
: \link[ggplot2:ggplot]{ggplot2::ggplot} of the plot and it saves the result to
\verb{product/plots/firing_qqplot_by_treatment_<experiment_tag>_<date_code>.pdf}
and
\verb{product/plots/firing_qqplot_by_treatment_<experiment_tag>_<date_code>.png}
It save both .pdf and .png because it's easier to email etc small pngs
while for use in an a manuscript having the vector version means that it`
can be tweaked with illustrator
}
\description{
Plot a grid of plots, with x-axis in microseconds and y-axis in voltage
One for each waveform in the waveform data.frame
}
