% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_normalize_label_data.R
\name{rm_time_off_baseline}
\alias{rm_time_off_baseline}
\title{Remove Time Off Baseline}
\usage{
rm_time_off_baseline(
  exp,
  baseline_cutoff,
  baseline_treatment_name = "Baseline"
)
}
\arguments{
\item{exp}{\link{meapr-experiment}, a \link{meapr} experiment object}

\item{baseline_cutoff}{\code{numeric} trim the given amount of the start of the
baseline}

\item{baseline_treatment_name}{\code{character} name of the baseline treatment}
}
\value{
The \link{meapr-experiment} that was passed in as \code{exp}, but with the
baseline condition trimmed.
}
\description{
Remove Time Off Baseline
}
